Add cross-origin resource sharing (CORS) suport to django server
This patch installs django-cors-headers package, which is used as
a middleware which adds the Access-Control-Allow-Origin http header
to every response from the server. Without this header, any code
hosted on a different website will be denied access to the server,
which meant that the frontend code always failed to fetch data from
the server unless you installed an unsafe plugin, which users of
the site would not want to do.