Add proof of concept code for google authentication
This uses the react-google-login package to do all of the hard work.
Once the user logs in through google, the app posts a message to the
server containing the oauth key and user information. It then prints
out the server's response.

This message to the server includes the a token, email, name,
and profilePicture (which is a link to they're google profile),
so this information can be stored with the user on the server.

The important fields are the access_token and refresh_token. The
access token needs to be included with all requests to the server
(at least those that modify data), and the refresh_token needs to
be used to refresh the login session after a short time period. This
most likely can be saved in the user's browser storage, so that they
wouldn't have to login each time.