Add and use methods for doing user tokenRefresh and login
The app now also saves the user's information in localStorage so that
the user will stay logged in even if they refresh the page.

For refreshing the user's access token, only the refresh_token field
is needed, but since refreshing the token doesn't give any user
information, additional user information must be stored (in this case
their email). This is probably a bad idea.