Allow editing and creating projects on the front end
This works by including a speciific HTTP header with the request.
'Authorization': 'Bearer {access token}'

There are still some issues with the project form though, since it
appears to completely disregard the member list. So editing a project
will clear its member list, and created projects will have no members
(they should have owner in the list). Because of this, I had to make
the ProjectModal handle the case where there are no members, but this
shouldn't ever occur if things worked properly.