Set up the logged in user before rendering any components
This will allow components to be able to check their props.user
without being temporarily lied to. When restoring the user from
localStorage was done in componentDidMount, there was a small time
period where the user was null, even though there actually was a
user. This would mean that the user couldn't safely be accessed in
a constructor or componentDidMount function, and attempts to redirect
the user if they weren't logged in would have false positives.