Fixed some warnings and errors
I removed the CardText element because it wasn't doing anything to the styling, and it was raising warnings due to the mixing of in-line and block elements. I also added a check to make sure that an error isn't raised if someone who is not logged in tries to view a profile page as if they're logged in.